requirements.txt
blinker==1.6.2
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
et-xmlfile==1.1.0
Flask==2.3.3
fonttools==4.42.1
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
Werkzeug==2.3.7
zipp==3.16.2
